# docs/openapi.yaml
openapi: 3.1.0
info:
  title: vet402 API
  version: 1.0.0
  description: |
    Independent verification for the x402 agent-payment economy. Scores payee
    wallets and ERC-8004 agents on Base, returns a 0-100 score with an
    ALLOW/WARN/BLOCK recommendation, and publishes what happened after those
    verdicts.

    **Disclaimer:** Scores are informational only and do not constitute a guarantee,
    credit assessment, or investment advice. Every response that carries a score or
    a published rate repeats this string in the payload, so the caveat travels with
    the number.

    **Rate limits.** Authenticated calls: a per-account monthly quota (Free 1,000
    / Pro 50,000 / Scale 500,000), shared across keys, reset at the UTC month
    boundary. Each `/score` is 1 unit; a `/scores/batch` of N agents is N units.
    Scored responses carry `X-RateLimit-Limit`, `X-RateLimit-Used`, and
    `X-RateLimit-Remaining`. There is no per-second burst throttle on
    authenticated calls — you may spend the quota as fast as you like; a `429`
    with `Retry-After` means the monthly quota is exhausted. Separate IP-based
    abuse throttles (e.g. 60 auth-failures/min/IP) exist only to blunt
    key-guessing and scraping and are not hit by valid authenticated traffic.

    Key-less calls are throttled per IP instead. Every response on those paths
    carries `RateLimit-Limit` — the ceiling for that path, true for every caller.
...

...
    response.

  contact:
    name: vet402 API Support
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/payees/verify:
    get:
      operationId: previewPayeeVerifyMessage
      tags: [Payees]
      summary: Preview the canonical message to sign for payee verification
      description: |
        Returns the exact message a payee wallet must sign to register via
        `POST /api/v1/payees/verify`. No API key; throttled to 30 requests per
        minute per IP like the other key-less paths. The message is four lines
        (wallet + name) or five when `url` is supplied — the URL is bound into
        the signature so a published proof cannot overwrite the profile link.
        Pass the same `url` query you will POST, or omit it for a name-only claim.
      security: []
      parameters:
        - name: wallet
          in: query
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: true
          schema: { type: string, maxLength: 80 }
        - name: url
          in: query
          required: false
          schema: { type: string, format: uri }
          description: Optional https profile URL; when set it is bound into the signed message.
      responses:
        "200":
          description: Canonical message to sign
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: verifyPayee
      tags: [Payees]
      summary: Register a verified payee by signing the canonical message
      description: |
        Free, no API key. Sign the message from `GET /api/v1/payees/verify` (or
        build it yourself per that endpoint's description) with the payee
        wallet — a valid EIP-191 signature (EIP-6492 for smart accounts) is
        the proof of control and is the main anti-abuse gate, backed by an
        8-per-minute-per-IP write throttle. Publishes a
        public profile at `/payee/{wallet}` and an embeddable SVG badge at
        `/api/badge/{wallet}.svg`. Verification proves wallet control only;
        trust scores stay computed independently.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet, name, signature]
              properties:
                wallet: { type: string }
                name: { type: string, maxLength: 80 }
                url: { type: string, format: uri, description: "Must be https://" }
                signature: { type: string, maxLength: 4000 }
      responses:
        "200":
          description: Payee verified and published
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  profile: { type: string }
                  badge: { type: string }
        "400":
          description: |
            Bad request — includes `invalid_wallet_address`, `url_must_be_https`,
            and `signature_mismatch` (whose response body also carries
            `expectedMessage`, so a rejected signature is self-diagnosing).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  expectedMessage:
                    type: string
                    description: Present only on signature_mismatch.
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          description: Store unavailable (schema not yet migrated, or DB unreachable).
...
  /api/v1/agents/verify:
    get:
      operationId: previewAgentVerifyMessage
      tags: [Agents]
      summary: Preview the canonical message to sign for an agent passport
      description: |
        Returns the exact message the agent's on-chain wallet must sign to
        register a passport, with that wallet resolved from the ERC-8004
        registry. Five lines, or six when `url` is supplied — the URL is bound
        into the signature. Pass the same `url` query you will POST, or omit
        it for a name-only claim. Key-less; 30 requests per minute per IP.
      security: []
      parameters:
        - name: agentId
          in: query
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: true
          schema: { type: string, maxLength: 80 }
        - name: url
          in: query
          required: false
          schema: { type: string, format: uri }
          description: Optional https profile URL; when set it is bound into the signed message.
      responses:
        "200":
          description: Canonical message to sign, with the resolved wallet
          content:
            application/json:
              schema:
                type: object
                required: [agentId, wallet, message]
                properties:
                  agentId: { type: string }
                  wallet: { type: string }
                  message: { type: string }
        "400":
          description: |
            `invalid_agent_id`, `invalid_name`, or `agent_wallet_unbound` — the
            last meaning the identity has no wallet bound on-chain, so no
            signature could prove control of it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          description: Chain unavailable — fail-closed, never a stored unverifiable claim.
    post:
      operationId: verifyAgent
      tags: [Agents]
      summary: Register an agent passport by signing the canonical message
      description: |
        Free, no API key. The on-chain wallet is resolved first, so the
        signature means "this agent", not merely "some wallet". Two throttles
        apply: 8 requests per minute per IP, and 4 per minute per agentId (a
        valid signature proves control, but one agent may not re-register
        without limit).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, name, signature]
              properties:
                agentId: { type: string }
                name: { type: string, maxLength: 80 }
                url: { type: string, format: uri, description: "Must be https://" }
                signature: { type: string, maxLength: 4000 }
      responses:
        "200":
          description: Passport registered and published
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  agentId: { type: string }
                  wallet: { type: string }
                  passport: { type: string }
                  profile: { type: string }
                  badge: { type: string }
        "400":
          description: |
            `invalid_request`, `invalid_agent_id`, `url_must_be_https`,
            `agent_wallet_unbound`, or `signature_mismatch` (whose body also
            carries `expectedMessage`, so a rejected signature is
            self-diagnosing).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  expectedMessage:
                    type: string
                    description: Present only on signature_mismatch.
        "429":
          description: |
            Throttled. `scope: "agent"` marks the per-agentId limit rather than
            the per-IP one.
        "503":
          description: Chain or store unavailable.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key issued from the developer dashboard
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: |
            Stable machine-readable error code. Request/body validation failures
            (malformed JSON, failed schema checks, empty/oversized batches, etc.)
            always use `invalid_request`.
          enum:
            - invalid_request
            - invalid_agent_id
            - invalid_wallet_address
            - invalid_tx_hash
            - attestation_unverifiable
            - missing_api_key
            - invalid_api_key
            - auth_unavailable
            - rate_limit_exceeded
            - plan_upgrade_required
            - scoring_unavailable
            - payment_ingest_unavailable
            # Key-less verification, watchlist, accuracy and the JSON 404
            # (2026-08-13: these paths were live but undocumented, so their
            # codes were missing from an enum a generated client validates on).
            - invalid_name
            - invalid_id
            - url_must_be_https
            - agent_wallet_unbound
            - signature_mismatch
            - chain_unavailable
            - store_unavailable
            - accuracy_unavailable
            - watchlist_unavailable
            - limit_reached
            - rate_limited
            - not_found
            - forbidden
        details:
          type: object
          description: Present for `invalid_request` — Zod validation error detail (flattened).
        reason:
          type: string
          description: Present for `attestation_unverifiable` — human-readable reason the on-chain check failed.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: |
        Rate limit exceeded. On authenticated paths this means the monthly quota
        is spent (`rate_limit_exceeded`); on key-less paths it is the per-IP
        window (`rate_limited`).
      headers:
        Retry-After:
          description: Seconds until the window resets.
          schema: { type: integer }
        RateLimit-Limit:
          description: Key-less paths — ceiling for the IP window. Sent on every response.
          schema: { type: integer }
        RateLimit-Remaining:
          description: >-
            Key-less paths — requests left in your current IP window. Per-caller,
            so it is sent here and on other uncached responses, but not on a
            CDN-cached 200.
          schema: { type: integer }
        RateLimit-Reset:
          description: >-
            Key-less paths — unix epoch seconds when your window resets. Same
            per-caller rule as RateLimit-Remaining.
          schema: { type: integer }
        X-RateLimit-Limit:
          description: Authenticated paths — monthly plan quota.
          schema: { type: integer }
        X-RateLimit-Used:
          description: Authenticated paths — quota consumed this month.
          schema: { type: integer }
        X-RateLimit-Remaining:
          description: Authenticated paths — quota left this month.
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: rate_limit_exceeded
              retryAfter:
                type: integer